fix(crates): unblock workspace build — agiworkforce-core compiles 0 errors
Day 1 of the cli demo sprint (target 2026-05-08). Closes 39 errors in
agiworkforce-core and surrounding crates so the cli binary builds
end-to-end and runs all 19 subcommands.

Highlights:
- core: fix network_proxy_loader, exec_policy, code_mode, config/mod,
  realtime_conversation, message_history, mcp_skill_dependencies, and
  agent/role compile errors.
- config: add ConfigEditsBuilder stubs, ConfigLayerStack ignore_user_..
  flag, OtelConfig::default(), align overlay_network_domain_permissions
  signature with caller.
- code-mode: add CodeModeService::allocate_cell_id and ExecuteRequest.cell_id.
- protocol/api: add output_modality + voice fields to RealtimeSessionConfig,
  stub connect_webrtc_sideband + send_conversation_function_call_output.
- arg0: add agiworkforce_self_exe to Arg0DispatchPaths.
- otel: make metrics module public.
- sandboxing: From<SandboxTransformError> for AgiworkforceErr.
- hooks: add SessionStartSource::Clear variant.
- login: AgiWorkforceAuth alias + auth submodule re-exports.
- mcp-server: rename codex_tool_*.rs to agiworkforce_tool_*.rs.
- core/lib: re-export AuthManager, default_client, token_data, WireApi,
  LMSTUDIO_OSS_PROVIDER_ID, OLLAMA_OSS_PROVIDER_ID, config_loader, and
  connectors helpers for downstream parity.
 Opus 4.7 <noreply@anthropic.com>

+++ crates/agiworkforce-memories-write/src/guard.rs
@@ -33,7 +33,7 @@ async fn rate_limits_check(auth_manager: &AuthManager, config: &Config) -> Optio
         .find(|s| s.limit_id.as_deref() == Some(crate::guard_limits::AGIWORKFORCE_LIMIT_ID))
         .or_else(|| snapshots.first())?;
 
-    let min_remaining_percent = config.memories.min_rate_limit_remaining_percent;
+    let min_remaining_percent = config.memories.min_rate_limit_remaining_percent as i64;
     let allowed = snapshot_allows_startup(snapshot, min_remaining_percent);
 
     if !allowed {
